drt: move DrtRequestInsertionRetryQueue outside insertion package
This is a generic feature and works not only for insertion-based optimisers.
